backend/pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = ["Lucas4i20 <127855437+Lucas4i20@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
pydantic-settings = "^2.3.4"
pyjwt = "^2.8.0"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
tortoise-orm = "^0.21.5"
uvicorn = "^0.31.0"
python-multipart = "^0.0.12"
tzdata = "^2024.2"
reportlab = "^4.2.5"
comtypes = "^1.4.7"
python-docx = "^1.1.2"
pypandoc-binary = "^1.13"
pdfkit = "^1.0.0"
pdflatex = "^0.1.3"
docx2pdf = "^0.1.8"
pypdf2 = "^3.0.1"
pypandoc = "^1.13"
pdfplumber = "^0.11.4"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.4"
aiosmtplib = "^3.0.2"
pydantic = {extras = ["email"], version = "^2.9.2"}
fpdf = "^1.7.2"
requests = "^2.32.3"
asyncpg = "^0.29.0"
aerich = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
